load image from rgb-d camera tested depth map

+++ src/utility/DebugUtility.cpp
@@ -8,7 +8,7 @@
 
 namespace cxy
 {
-    void DebugUtility::DisplayImage(int width, int height, int type, void* data, std::string windowName)
+    void DebugUtility::DisplayImage(int width, int height, int type, void* data, std::string windowName, bool keep)
     {
         /*
              * Debug: display the image
@@ -18,7 +18,24 @@ namespace cxy
         imageTest.convertTo(imageTest2, CV_8U);
         cv::imshow(windowName, imageTest2);
         cv::waitKey(0);
-        cv::destroyWindow(windowName);
+
+        if ( ! keep)
+            cv::destroyWindow(windowName);
+        return ;
+    }
+
+    void DebugUtility::DisplayImage(cv::Mat image, std::string windowName, bool keep)
+    {
+        /*
+             * Debug: display the image
+             */
+        cv::Mat imageTest2;
+        image.convertTo(imageTest2, CV_8U);
+        cv::imshow(windowName, imageTest2);
+        cv::waitKey(0);
+
+        if ( ! keep)
+            cv::destroyWindow(windowName);
         return ;
     }
